Move BoundingBox and summaries to Physics target
