+++ src/NBody/Physics/CMakeLists.txt
@@ -12,6 +12,13 @@ add_library(Physics
         Position.h
         Velocity.h
         Mass.cpp Mass.h
+        BoundingBox.h
+        SummaryType.h
+        Summaries/CenterOfMassSummary.h
+        Summaries/QuadrupoleMassSummary.h
+        Summaries/AccelerationSummary.h
+        Summaries/DualSummary.h
+        Summaries/BoundingBoxSummary.h
         Rule.h Force.h Acceleration.h Quadrupole.h)
 
 target_include_directories(Physics PUBLIC
