+++ src/NBody/Physics/Summaries/BoundingBoxSummary.h
@@ -5,7 +5,7 @@
 #ifndef N_BODY_BOUNDINGBOXSUMMARY_H
 #define N_BODY_BOUNDINGBOXSUMMARY_H
 
-#include <NBody/Simulation/Solvers/Trees/SummaryType.h>
+#include "NBody/Physics/SummaryType.h"
 
 namespace NBody {
